+++ monitoring_app/src/fi/fullerine/greenhouse/shared/SensorData.java
@@ -1,40 +1,38 @@
 package fi.fullerine.greenhouse.shared;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
-import java.math.BigInteger;
 
 public class SensorData implements Serializable {
 	private static final long serialVersionUID = 1L;
 
-	private BigDecimal temperature;
-	private BigDecimal humidity;
-	private BigInteger moisture1;
-	private BigInteger moisture2;
+	private Double temperature;
+	private Double humidity;
+	private Integer moisture1;
+	private Integer moisture2;
 	private Boolean waterBarrelEmpty;
 	
-	public BigDecimal getTemperature() {
+	public Double getTemperature() {
 		return temperature;
 	}
-	public void setTemperature(BigDecimal temperature) {
+	public void setTemperature(Double temperature) {
 		this.temperature = temperature;
 	}
-	public BigDecimal getHumidity() {
+	public Double getHumidity() {
 		return humidity;
 	}
-	public void setHumidity(BigDecimal humidity) {
+	public void setHumidity(Double humidity) {
 		this.humidity = humidity;
 	}
-	public BigInteger getMoisture1() {
+	public Integer getMoisture1() {
 		return moisture1;
 	}
-	public void setMoisture1(BigInteger moisture1) {
+	public void setMoisture1(Integer moisture1) {
 		this.moisture1 = moisture1;
 	}
-	public BigInteger getMoisture2() {
+	public Integer getMoisture2() {
 		return moisture2;
 	}
-	public void setMoisture2(BigInteger moisture2) {
+	public void setMoisture2(Integer moisture2) {
 		this.moisture2 = moisture2;
 	}
 	public Boolean getWaterBarrelEmpty() {
@@ -43,6 +41,5 @@ public class SensorData implements Serializable {
 	public void setWaterBarrelEmpty(Boolean waterBarrelEmpty) {
 		this.waterBarrelEmpty = waterBarrelEmpty;
 	}
-	
 	
 }
